git-remote-gitern: Use git pull --rebase when update mirror repo
Mirror repo is just that - it should mirror remote exactly. If not use
--rebase, if remote was e.g. rebased, a stray merge may be created in
the mirror, not correspondingto anything in remote and causing confusion.

Note that it would be better to use fetch +ref:ref for each ref, but
that's extra processing and won't work still due to mirror being non-bare
repo.

pull --rebase tho may leave some stray commits in the mirror branch (but
at least they don't seem to interfere).

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>